16-bit arithmetic, branch relaxation, and full climber game
- 16-bit multiply: conditional ASL/ROL chain when followed by Conv_u2
- 16-bit add/sub: lo/hi byte decomposition with carry/borrow propagation
- 16-bit shift right: TXA for >>8, LSR TEMP/ROR A for >>N
- Conv_u1 clears _ushortInAX flag when truncating to byte
- WriteLdc bypass: keeps A:X for binary ops instead of pushax
- Branch relaxation: auto-replace out-of-range BNE/BEQ with JMP trampolines
- Zerobss supports >255 local bytes (pages + remainder)
- locals type widened from byte to ushort throughout transpiler chain
- Full climber game: 685 lines, platform game with scrolling, enemies,
  collision detection, BCD scoring, floor generation, metasprites
- 3 new RoslynTests: Ushort16BitMultiply, Ushort16BitAdd, UshortStoreAndLoad

+++ src/dotnes.tests/TranspilerTests.cs
@@ -107,7 +107,7 @@ public class TranspilerTests
         using var transpiler = new Transpiler(dll, [], _logger);
         
         // Build using single-pass transpilation
-        var program = transpiler.BuildProgram6502(out ushort sizeOfMain, out byte locals);
+        var program = transpiler.BuildProgram6502(out ushort sizeOfMain, out ushort locals);
 
         // Verify the program has blocks
         Assert.True(program.BlockCount > 0, "Program should have blocks");
